fixed signed vs unsigned compiler warning

+++ main.cpp
@@ -23,7 +23,7 @@ struct DrawOperation
   static std::vector<DrawOperation *> drawQueue;
   static void Draw(Magick::Image &canvas)
   {
-    for(int i = 0; i < drawQueue.size(); i++)
+    for(size_t i = 0; i < drawQueue.size(); i++)
     {
       DrawOperation *pDraw = drawQueue[i];
       canvas.composite(*(pDraw->pImage), pDraw->x, pDraw->y);
